requirements.txt
discord.py>=2.3.2
asyncpg>=0.29.0
SQLAlchemy>=2.0.23
APScheduler>=3.10.4
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

